Per Slony-I bug 41 http://www.slony.info/bugzilla/show_bug.cgi?id=41
The xreflabel tags being generated by autodoc are concatenating
together the schema name and the function name.

If I put a "." in between, this short-circuits it, which seems to do
roughly the right thing.  I was expecting to get things like
"schemadoc.my_function(a,b)", and instead got "myfunction(a,b)", but
I actually prefer the latter :-).